# docs/swagger.yaml
basePath: /api
definitions:
...
  github_com_fisker_zjump-backend_internal_model.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@keyops.com
    name: API Support
  description: KeyOps 基础设施管理平台 API 文档
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: KeyOps API
  version: "2.0"
paths:
...
  /api/k8s/search:
    get:
      consumes:
      - application/json
      parameters:
      - description: 搜索关键词
        in: query
        name: q
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_fisker_zjump-backend_internal_model.Response'
            - properties:
                data:
                  additionalProperties: true
                  type: object
              type: object
      security:
      - BearerAuth: []
      summary: 全局搜索
      tags:
      - K8s Search
...
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"